all change

+++ XRFOC_Lib/include/mc_pid.h
@@ -5,16 +5,12 @@
  *  @file       XRFOC_Lib/include/mc_pid.h
  *  @brief      Pid algorithm implemented.
  *  @author:    RunDong7582
- *  @date       2025 3/20 15:59 -> 2025 4/1 16:26
- *  @version    XRFOC v0.2
+ *  @date       2025.5.19 18:32
+ *  @version    XRFOC release version 
  */
 
 /*--------inc--------------*/ 
 #include "../common/mc_common.h"
-#include <stdint.h>
-
-/*--------variable---------*/
-// volatile uint32_t timer_counter = 0;
 
 /*--------pid struct------*/
 typedef struct {
@@ -27,8 +23,8 @@ typedef struct {
     float error_prev;
     float output_prev;
     float integral_prev;
-    unsigned long timestamp_prev;
-
+    // unsigned long timestamp_prev;
+    uint32_t timestamp_prev;
 } PIDController;
 
 float pid_operator ( PIDController* pid, float error );
